Move launch related helpers into launch_util module
