+++ integration/experiment/Makefile.mk
@@ -31,6 +31,7 @@
 
 EXTRA_DIST += integration/experiment/common_args.py \
               integration/experiment/__init__.py \
+              integration/experiment/launch_util.py \
               integration/experiment/machine.py \
               integration/experiment/util.py \
               # end
