+++ integration/experiment/util.py
@@ -39,7 +39,6 @@ import subprocess
 import io
 import yaml
 import shlex
-import time
 
 import geopmpy.launcher
 
@@ -129,114 +128,4 @@ def geopmread(read_str):
             result = float(last_line)
         except ValueError:
             result = yaml.safe_load(output).values()[0]
-    return result
-
-
-def launch_run(agent_conf, app_conf, run_id, output_dir, extra_cli_args,
-               num_nodes, redirect_stdout=True):
-    # launcher and app should create files in output_dir
-    start_dir = os.getcwd()
-    os.chdir(output_dir)
-
-    # TODO: why does launcher strip off first arg, rather than geopmlaunch main?
-    argv = ['dummy', detect_launcher()]
-
-    agent_name = 'monitor'
-    if agent_conf is not None:
-        agent_name = agent_conf.get_agent()
-        agent_conf.write()
-    if agent_name != 'monitor':
-        argv.append('--geopm-agent=' + agent_conf.get_agent())
-        argv.append('--geopm-policy=' + agent_conf.get_path())
-
-    app_name = app_conf.name()
-
-    uid = '{}_{}_{}'.format(app_name.lower(), agent_name, run_id)
-    report_path = '{}.report'.format(uid)
-    trace_path = '{}.trace'.format(uid)
-    profile_name = uid
-    log_path = '{}.log'.format(uid)
-
-    # TODO: these are not passed to launcher create()
-    # some are generic enough they could be, though
-    extra_cli_args += ['--geopm-report', report_path,
-                       '--geopm-trace', trace_path,
-                       '--geopm-profile', profile_name,
-    ]
-    argv.extend(extra_cli_args)
-
-    # extra geopm args needed by app
-    argv.extend(app_conf.get_custom_geopm_args())
-
-    argv.extend(['--'])
-
-    bash_path = app_conf.make_bash(uid)
-    argv.extend([bash_path])
-
-    num_ranks = app_conf.get_rank_per_node() * num_nodes
-
-    launcher = geopmpy.launcher.Factory().create(argv,
-                                                 num_node=num_nodes,
-                                                 num_rank=num_ranks)
-    if redirect_stdout:
-        with open(log_path, 'w') as outfile:
-            # TODO: some apps print to stderr. use stderr=outfile
-            launcher.run(stdout=outfile)
-    else:
-        launcher.run()
-
-    # Get app-reported figure of merit
-    fom = app_conf.parse_fom(log_path)
-    # Append to report
-    with open(report_path, 'a') as report:
-        report.write('\nFigure of Merit: {}\n'.format(fom))
-
-    # return to previous directory
-    os.chdir(start_dir)
-
-
-class LaunchConfig():
-    def __init__(self, app_conf, agent_conf, name):
-        self._app_conf = app_conf
-        self._agent_conf = agent_conf
-        self._name = name
-
-    def app_conf(self):
-        return self._app_conf
-
-    def agent_conf(self):
-        return self._agent_conf
-
-    def run_id(self, iteration):
-        return '{}_{}'.format(self._name, iteration)
-
-
-def launch_all_runs(targets, num_nodes, iterations, extra_cli_args, output_dir, cool_off_time=60):
-    '''
-    targets: a list of LaunchConfig
-    iteration: integer number of iterations
-    '''
-    if not os.path.exists(os.path.join(output_dir, 'machine.json')):
-        raise RuntimeError('Missing machine file in {}; call machine.init_output_dir() before calling this function.'.format(output_dir))
-
-    for iteration in range(iterations):
-        for tar in targets:
-            agent_conf = tar.agent_conf()
-            app_conf = tar.app_conf()
-            run_id = tar.run_id(iteration)
-
-            launch_run(agent_conf, app_conf, run_id, output_dir,
-                       extra_cli_args=extra_cli_args,
-                       num_nodes=num_nodes)
-
-            # rest to cool off between runs
-            time.sleep(cool_off_time)
-
-
-def geopm_signal_args(report_signals, trace_signals):
-    result = []
-    if report_signals is not None:
-        result += ['--geopm-report-signals=' + ','.join(report_signals)]
-    if trace_signals is not None:
-        result += ['--geopm-trace-signals=' + ','.join(trace_signals)]
     return result
